+++ haskell/src/Lorentz/Contracts/StablecoinFA1_2/Types.hs
@@ -6,23 +6,14 @@
 module Lorentz.Contracts.StablecoinFA1_2.Types
   ( Storage(..)
   , Parameter(..)
-  , stablecoinFA1_2Contract
-
-  , GetCounterParam
-  , GetDefaultExpiryParam
   ) where
 
-import Fmt (pretty)
 
 import Lorentz as L
 import Lorentz.Contracts.Spec.ApprovableLedgerInterface qualified as AL
 import Lorentz.Contracts.Spec.TZIP16Interface (MetadataMap)
-import Morley.Michelson.Parser.Types (MichelsonSource(MSFile))
-import Morley.Michelson.Untyped qualified as U
-import Test.Cleveland.Michelson.Import (embedTextFile, readUntypedContract)
 
 import Lorentz.Contracts.Stablecoin qualified as S
-import Lorentz.Contracts.StablecoinPath (stablecoinFA1_2Path)
 
 data Storage = Storage
   { sDefaultExpiry :: S.Expiry
@@ -38,9 +29,20 @@ data Storage = Storage
   , sTransferlistContract :: Maybe Address
   }
 
+customGeneric "Storage" ligoLayout
 deriving anyclass instance IsoValue Storage
 deriving anyclass instance HasAnnotation Storage
-customGeneric "Storage" ligoLayout
+
+-- | Similar to 'AL.Parameter' from @morley-ledgers@, but with a layout compatible with
+-- the ligo implementation
+data FA1_2Parameter
+  = Transfer       AL.TransferParams
+  | Approve        AL.ApproveParams
+  | GetAllowance   AL.GetAllowanceArg
+  | GetBalance     AL.GetBalanceArg
+  | GetTotalSupply AL.GetTotalSupplyArg
+customGeneric "FA1_2Parameter" ligoLayout
+deriving anyclass instance IsoValue FA1_2Parameter
 
 data Parameter
   = Accept_ownership
@@ -48,8 +50,6 @@ data Parameter
   | Change_master_minter S.ChangeMasterMinterParam
   | Change_pauser S.ChangePauserParam
   | Configure_minter S.ConfigureMinterParam
-  | Get_counter GetCounterParam
-  | Get_default_expiry GetDefaultExpiryParam
   | Mint S.MintParams
   | Pause
   | Permit S.PermitParam
@@ -58,24 +58,10 @@ data Parameter
   | Set_transferlist S.SetTransferlistParam
   | Transfer_ownership S.TransferOwnershipParam
   | Unpause
-  | Call_FA1_2 AL.Parameter
-  deriving stock (Generic)
-  deriving anyclass (IsoValue)
+  | Call_FA1_2 FA1_2Parameter
+
+customGeneric "Parameter" ligoLayout
+deriving anyclass instance IsoValue Parameter
 
 instance ParameterHasEntrypoints Parameter where
   type ParameterEntrypointsDerivation Parameter = EpdRecursive
-
-type GetCounterParam = View_ () Natural
-
-type GetDefaultExpiryParam = View_ () S.Expiry
-
--- | Parse the metadata registry contract.
-stablecoinFA1_2Contract :: U.Contract
-stablecoinFA1_2Contract = $$(do
-  text <- embedTextFile stablecoinFA1_2Path
-  case readUntypedContract (MSFile stablecoinFA1_2Path) text of
-    Left err -> fail $ pretty err
-    -- Note: it's ok to use `error` here, because we just proved that the contract
-    -- can be parsed+typechecked.
-    Right _ -> [|| either (error . pretty) id $ readUntypedContract (MSFile stablecoinFA1_2Path) text ||]
-    )